edit javadoc and move tests
